menu components done
