+++ app/src/main/java/reservant_mobile/ui/components/Components.kt
@@ -1,7 +1,6 @@
 package reservant_mobile.ui.components
 
 import android.content.Context
-import android.net.Uri
 import android.widget.Toast
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.animateDp
@@ -116,14 +115,12 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.composed
-import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.drawBehind
 import androidx.compose.ui.focus.onFocusChanged
 import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.ImageBitmap
 import androidx.compose.ui.graphics.RectangleShape
 import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.graphics.vector.ImageVector
@@ -131,9 +128,7 @@ import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.SpanStyle
 import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.buildAnnotatedString
 import androidx.compose.ui.text.font.FontStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.input.KeyboardType
@@ -155,7 +150,6 @@ import org.osmdroid.util.GeoPoint
 import org.osmdroid.views.MapView
 import reservant_mobile.data.constants.Roles
 import reservant_mobile.data.models.dtos.OrderDTO
-import reservant_mobile.data.models.dtos.RestaurantMenuDTO
 import reservant_mobile.data.models.dtos.RestaurantMenuItemDTO
 import reservant_mobile.data.models.dtos.fields.FormField
 import reservant_mobile.data.services.UserService
@@ -262,6 +256,66 @@ fun OutLinedDropdownMenu(
         }
     }
 }
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun ComboBox(
+    expanded: MutableState<Boolean>,
+    value: String,
+    onValueChange: (String) -> Unit,
+    options: List<String>,
+    label: String,
+    isError: Boolean = false,
+    errorText: String = ""
+){
+
+    val onDismiss = { expanded.value = false }
+    var beginValidation by remember {
+        mutableStateOf(false)
+    }
+
+    ExposedDropdownMenuBox(
+        expanded = expanded.value,
+        onExpandedChange = {
+            expanded.value = !expanded.value
+            beginValidation = true
+        }
+    ) {
+        Column {
+            OutlinedTextField(
+                modifier = Modifier
+                    .padding(vertical = 8.dp)
+                    .menuAnchor(),
+                label = { Text(text = label) },
+                value = value,
+                onValueChange = {},
+                readOnly = true,
+                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded.value) },
+                shape = RoundedCornerShape(8.dp),
+                isError = isError && beginValidation
+            )
+
+            if (isError && beginValidation) Text(text = errorText, color = MaterialTheme.colorScheme.error)
+        }
+
+        ExposedDropdownMenu(
+            modifier = Modifier.exposedDropdownSize(matchTextFieldWidth = false),
+            expanded = expanded.value,
+            onDismissRequest = onDismiss
+        ) {
+            options.forEach {
+                DropdownMenuItem(
+                    text = { Text(text = it) },
+                    onClick = {
+                        onValueChange(it)
+                        onDismiss()
+                    }
+                )
+            }
+        }
+    }
+}
+
 
 @Composable
 fun ButtonComponent(
@@ -516,304 +570,6 @@ fun Modifier.shimmer(): Modifier = composed {
         .onGloballyPositioned {
             size = it.size
         }
-}
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun ComboBox(
-    expanded: MutableState<Boolean>,
-    value: String,
-    onValueChange: (String) -> Unit,
-    options: List<String>,
-    label: String,
-    isError: Boolean = false,
-    errorText: String = ""
-){
-
-    val onDismiss = { expanded.value = false }
-    var beginValidation by remember {
-        mutableStateOf(false)
-    }
-
-    ExposedDropdownMenuBox(
-        expanded = expanded.value,
-        onExpandedChange = {
-            expanded.value = !expanded.value
-            beginValidation = true
-        }
-    ) {
-        Column {
-            OutlinedTextField(
-                modifier = Modifier
-                    .padding(vertical = 8.dp)
-                    .menuAnchor(),
-                label = { Text(text = label) },
-                value = value,
-                onValueChange = {},
-                readOnly = true,
-                trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = expanded.value) },
-                shape = RoundedCornerShape(8.dp),
-                isError = isError && beginValidation
-            )
-
-            if (isError && beginValidation) Text(text = errorText, color = MaterialTheme.colorScheme.error)
-        }
-
-        ExposedDropdownMenu(
-            modifier = Modifier.exposedDropdownSize(matchTextFieldWidth = false),
-            expanded = expanded.value,
-            onDismissRequest = onDismiss
-        ) {
-            options.forEach {
-                DropdownMenuItem(
-                    text = { Text(text = it) },
-                    onClick = {
-                        onValueChange(it)
-                        onDismiss()
-                    }
-                )
-            }
-        }
-    }
-}
-
-@Composable
-fun MenuCard(
-    showConfirmDeletePopup: MutableState<Boolean> = mutableStateOf(false),
-    showEditPopup: MutableState<Boolean> = mutableStateOf(false),
-    name: FormField,
-    altName: FormField,
-    menuType: FormField,
-    menuTypes: List<String>,
-    dateFrom: FormField,
-    dateUntil: FormField,
-    photo: ImageBitmap? = null,
-    menu: RestaurantMenuDTO,
-    onFilePicked: (Uri?) -> Unit,
-    onEditClick: () -> Unit,
-    onDeleteClick: () -> Unit,
-    clearFields: () -> Unit,
-    onClick: () -> Unit,
-    isFetching: Boolean = false,
-    isSaving: Boolean = false,
-    fileTooLarge: Int = -1,
-    fileErrors: Int = -1,
-    isNameInvalid: Boolean = false,
-    isAltNameInvalid: Boolean = false,
-    isMenuTypeInvalid: Boolean = false
-) {
-
-    when {
-        showConfirmDeletePopup.value -> {
-            CountDownPopup(
-                icon = Icons.Filled.DeleteForever,
-                title = stringResource(id = R.string.confirm_delete_title),
-                text = stringResource(id = R.string.confirm_delete_text),
-                onConfirm = {
-                    onDeleteClick()
-                },
-                onDismissRequest = {showConfirmDeletePopup.value = false},
-                confirmText = stringResource(id = R.string.label_yes_capital),
-                dismissText = stringResource(id = R.string.label_cancel),
-                isSaving = isSaving
-            )
-        }
-
-        showEditPopup.value -> {
-            name.value = menu.name
-            altName.value = menu.alternateName ?: ""
-            menuType.value = menu.menuType
-            dateFrom.value = menu.dateFrom
-            dateUntil.value = menu.dateUntil ?: ""
-
-            MenuPopup(
-                title = { Text(text = stringResource(id = R.string.label_edit_menu)) },
-                hide = {showEditPopup.value = false},
-                onConfirm = onEditClick,
-                clear = clearFields,
-                onFilePicked = onFilePicked,
-                name = name,
-                altName = altName,
-                menuType = menuType,
-                menuTypes = menuTypes,
-                dateFrom = dateFrom,
-                dateUntil = dateUntil,
-                isSaving = isSaving,
-                fileTooLarge = fileTooLarge,
-                fileErrors = fileErrors,
-                isNameInvalid = isNameInvalid,
-                isAltNameInvalid = isAltNameInvalid,
-                isMenuTypeInvalid = isMenuTypeInvalid
-            )
-        }
-    }
-    
-
-    val loadingModifier = when {
-        isFetching -> Modifier
-            .shimmer()
-            .alpha(0F)
-        else -> Modifier
-    }
-
-    Card(
-        elevation = CardDefaults.cardElevation(8.dp),
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp)
-            .clickable(onClick = onClick)
-            .then(loadingModifier)
-    ) {
-        Column {
-            if (photo != null){
-                Image(
-                    bitmap = photo,
-                    contentDescription = "${menu.name}_photo",
-                    modifier = Modifier.fillMaxWidth()
-                )
-            } else {
-                Image(
-                    painterResource(id = R.drawable.unknown_image),
-                    contentDescription = "placeholder_photo",
-                    modifier = Modifier
-                        .size(140.dp)
-                        .fillMaxWidth()
-                        .align(Alignment.CenterHorizontally)
-                )
-            }
-
-
-
-            Box(
-                modifier = Modifier.fillMaxWidth()
-            ) {
-
-                Column(
-                    modifier = Modifier
-                        .align(Alignment.TopStart)
-                ) {
-
-                    val namePadding = when {
-                        menu.alternateName == null -> 8.dp
-                        else -> 2.dp
-                    }
-
-                    Text(
-                        text = menu.name,
-                        style = MaterialTheme.typography.titleMedium.copy(fontSize = 20.sp),
-                        modifier = Modifier
-                            .padding(start = 8.dp, end = 8.dp, bottom = namePadding, top = 8.dp)
-                    )
-
-                    menu.alternateName?.let {
-                        Text(
-                            text = it,
-                            style = MaterialTheme.typography.titleSmall,
-                            fontWeight = FontWeight.Light,
-                            modifier = Modifier
-                                .padding(start = 8.dp, end = 8.dp, bottom = 8.dp)
-                        )
-                    }
-
-                    menu.dateUntil?.let {
-                        Text(
-                            text = buildAnnotatedString {
-                                append(stringResource(id = R.string.label_limited_time))
-                                append(": ")
-                                pushStyle(SpanStyle(fontWeight = FontWeight.Bold))
-                                append(it)
-                            },
-                            style = MaterialTheme.typography.bodyMedium,
-                            modifier = Modifier
-                                .padding(start = 8.dp, end = 8.dp, bottom = 8.dp)
-                        )
-                    }
-                }
-
-                Row(
-                    modifier = Modifier
-                        .align(Alignment.BottomEnd)
-                ) {
-                    val buttonModifier = Modifier
-                        .align(Alignment.Bottom)
-                        .size(50.dp)
-                        .padding(6.dp)
-
-                    SecondaryButton(
-                        modifier = buttonModifier,
-                        onClick = { showEditPopup.value = true },
-                        imageVector = Icons.Filled.Edit,
-                        contentDescription = "EditMenuItem"
-                    )
-
-                    SecondaryButton(
-                        modifier = buttonModifier,
-                        onClick = { showConfirmDeletePopup.value = true },
-                        imageVector = Icons.Filled.DeleteForever,
-                        contentDescription = "delete"
-                    )
-                }
-
-                /*Text(
-                    text = menu.menuType,
-                    style = MaterialTheme.typography.bodyMedium,
-                    modifier = Modifier
-                        .align(Alignment.End)
-                        .padding(8.dp)
-                )*/
-
-            }
-        }
-    }
-}
-
-@Composable
-fun AddMenuButton(
-    name: FormField,
-    altName: FormField,
-    menuType: FormField,
-    menuTypes: List<String>,
-    dateFrom: FormField,
-    dateUntil: FormField,
-    onFilePicked: (Uri?) -> Unit,
-    clearFields: () -> Unit,
-    addMenu: () -> Unit,
-    isSaving: Boolean = false,
-    fileTooLarge: Int = -1,
-    fileErrors: Int = -1,
-    isNameInvalid: Boolean = false,
-    isAltNameInvalid: Boolean = false,
-    isMenuTypeInvalid: Boolean = false,
-    showAddDialog: MutableState<Boolean> = mutableStateOf(false)
-) {
-    when {
-        showAddDialog.value -> {
-            MenuPopup(
-                title = { Text(text = stringResource(id = R.string.label_add_menu)) },
-                hide = { showAddDialog.value = false },
-                onConfirm = addMenu,
-                clear = clearFields,
-                onFilePicked = onFilePicked,
-                fileTooLarge = fileTooLarge,
-                fileErrors = fileErrors,
-                name = name,
-                altName = altName,
-                menuType = menuType,
-                menuTypes = menuTypes,
-                dateFrom = dateFrom,
-                dateUntil = dateUntil,
-                isSaving = isSaving,
-                isNameInvalid = isNameInvalid,
-                isAltNameInvalid = isAltNameInvalid,
-                isMenuTypeInvalid = isMenuTypeInvalid
-            )
-        }
-    }
-
-    MyFloatingActionButton(
-        onClick = { showAddDialog.value = true }
-    )
 }
